requirements.txt
openai-whisper
pyaudio
pynput
keyboard
argparse
rumps
numpy
boto3
librosa

amazon-transcribe

pyperclip
